Move openapi3 decorators in namespace
