+++ packages/openapi3/package.json
@@ -17,22 +17,22 @@
     "cadl"
   ],
   "type": "module",
-  "main": "dist/src/openapi.js",
+  "main": "dist/src/index.js",
   "exports": {
-    ".": "./dist/src/openapi.js",
+    ".": "./dist/src/index.js",
     "./testing": "./dist/src/testing/index.js"
   },
   "typesVersions": {
     "*": {
       "*": [
-        "./dist/src/openapi.d.ts"
+        "./dist/src/index.d.ts"
       ],
       "testing": [
         "./dist/src/testing/index.d.ts"
       ]
     }
   },
-  "cadlMain": "dist/src/openapi.js",
+  "cadlMain": "dist/src/index.js",
   "engines": {
     "node": ">=14.0.0"
   },
