+++ packages/openapi3/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./decorators.js";
+export * from "./openapi.js";
